Use `std::cout` instead of `qInfo`
`qInfo` is pretty handy, but does include a number of features we
don't need, and `std::cout` is roughly twice as fast.

Also swapped `qSort` for `std::sort` as `qSort` is now obsoleted,
with `std::sort` the recommended replacement -
https://doc.qt.io/qt-5/qtalgorithms-obsolete.html#qSort

Switching to `std::sort` made no appreciable difference in
performance on my machine.